# go-impl/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for tracking collections",
        "title": "Collection API",
        "contact": {
            "name": "Robb Currall",
            "email": "rlcurrall@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/v1/comics": {
            "get": {
                "description": "Get all comics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comics"
                ],
                "summary": "Get all comics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owner of the comics",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Title of the comic",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Order by price",
                        "name": "order[price]",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Comic"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Add a new comic",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comics"
                ],
                "summary": "Add a new comic",
                "parameters": [
                    {
                        "description": "Add a comic",
                        "name": "comic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewComic"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Comic"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/comics/{id}": {
            "get": {
                "description": "Get comic by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comics"
                ],
                "summary": "Get comic by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comic ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Comic"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Remove comic by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comics"
                ],
                "summary": "Remove comic by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comic ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update comic by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comics"
                ],
                "summary": "Update comic by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comic ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Add a comic",
                        "name": "comic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewComic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Comic"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "An error occurred"
                }
            }
        },
        "models.Comic": {
            "type": "object",
            "properties": {
                "coverYear": {
                    "type": "string",
                    "example": "2011-02-03"
                },
                "created_at": {
                    "type": "string"
                },
                "genre": {
                    "type": "string",
                    "example": "superhero"
                },
                "grade": {
                    "type": "number",
                    "example": 9.2
                },
                "id": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string",
                    "example": "https://www.dccomics.com/sites/default/files/styles/covers192x291/public/comic-covers/2018/06/batman_v2_52_5b2ad8fd533996.66775403.jpg"
                },
                "issueNumber": {
                    "type": "string",
                    "example": "52"
                },
                "mainCharacter": {
                    "type": "string",
                    "example": "Batman"
                },
                "price": {
                    "type": "number",
                    "example": 32.21
                },
                "publisher": {
                    "type": "string",
                    "example": "DC"
                },
                "title": {
                    "type": "string",
                    "example": "Batman #52 (2011)"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "example": "cmx_collector42"
                }
            }
        },
        "models.NewComic": {
            "type": "object",
            "properties": {
                "coverYear": {
                    "type": "string",
                    "example": "2011-02-03"
                },
                "genre": {
                    "type": "string",
                    "example": "superhero"
                },
                "grade": {
                    "type": "number",
                    "example": 9.2
                },
                "imageUrl": {
                    "type": "string",
                    "example": "https://www.dccomics.com/sites/default/files/styles/covers192x291/public/comic-covers/2018/06/batman_v2_52_5b2ad8fd533996.66775403.jpg"
                },
                "issueNumber": {
                    "type": "string",
                    "example": "52"
                },
                "mainCharacter": {
                    "type": "string",
                    "example": "Batman"
                },
                "price": {
                    "type": "number",
                    "example": 32.21
                },
                "publisher": {
                    "type": "string",
                    "example": "DC"
                },
                "title": {
                    "type": "string",
                    "example": "Batman #52 (2011)"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}